backend/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
aiomysql
PyMySQL
python-jose[cryptography]
passlib[bcrypt]
pydantic[email]
python-multipart
reportlab
pydantic-settings
pytest
pytest-asyncio
httpx
bcrypt
aiosqlite
annotated-doc
annotated-types
anyio
certifi
cffi
charset-normalizer
click
colorama
cryptography
dnspython
ecdsa
email-validator
fastapi-cli
fastapi-cloud-cli
git-filter-repo
greenlet
h11
httpcore
httptools
idna
iniconfig
itsdangerous
Jinja2
markdown-it-py
MarkupSafe
mdurl
mysql-connector-python
orjson
packaging
pillow
pluggy
pyasn1
pycparser
pydantic_core
pydantic-extra-types
Pygments
python-dotenv
PyYAML
rich
rich-toolkit
rignore
rsa
ruff
sentry-sdk
shellingham
six
sniffio
starlette
typer
typing_extensions
typing-inspection
ujson
urllib3
watchfiles
websockets
pandas
openpyxl
alembic
slowapi
python-magic-bin